Refactor set_admins into configuration.py and update deployment params with nsg_config
